(508935) Handle missing SOTA directory in Dispatcher during SOTA

+++ inbm/dispatcher-agent/dispatcher/packagemanager/local_repo.py
@@ -131,8 +131,15 @@ class DirectoryRepo(IRepo):  # pragma: no cover
         """
         repo_path = self.get_repo_path()
         logger.debug("Erasing files directly underneath " + repo_path)
-        for the_file in os.listdir(repo_path):
-            file_path = os.path.join(self.get_repo_path(), the_file)
-            if os.path.isfile(file_path):
-                logger.debug("... unlinking: " + file_path)
-                os.unlink(file_path)
+        if not os.path.exists(repo_path):
+            raise DispatcherException(f"cannot delete files from {repo_path}: does not exist")
+        if not os.path.isdir(repo_path):
+            raise DispatcherException(f"cannot delete files from {repo_path}: is not a directory")
+        try:
+            for the_file in os.listdir(repo_path):
+                file_path = os.path.join(self.get_repo_path(), the_file)
+                if os.path.isfile(file_path):
+                    logger.debug("... unlinking: " + file_path)
+                    os.unlink(file_path)
+        except OSError as e:
+            raise DispatcherException("got OSError while deleting all in {repo_path}: {e}") from e
